backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
google-generativeai==0.8.3
anthropic==0.42.0
httpx==0.28.1
sse-starlette==2.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
